TP2/requirements.txt
python-dotenv>=1.0.0
pinecone>=5.0.0
groq>=0.11.0
sentence-transformers>=3.0.0
torch>=2.0.0
